tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-iteration-count_inline.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-iteration-count and inline element</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-iteration-count-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="animation-iteration-count applies to inline
   elements">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes fade {
         from {
            opacity: 1;
         }
         to {
            opacity: 0;
         }
      }
      span {
         color: blue;
         animation-duration: 700ms;
         animation-iteration-count: 2;
      }
      .test {
         animation-name: fade;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var span = document.querySelector('span');
         var log = document.querySelectorAll('p')[1].firstChild;
         function aboutEqual(observed, actual) {
            var maxDeviation = 0.1;
            return (Math.abs(observed - actual) < maxDeviation);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 2000);
         span.addEventListener('animationend', function(evt) {
            clearTimeout(timeout);
            var passed = aboutEqual(evt.elapsedTime, 1.4);
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         span.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue word below should fade out and reappear twice.</p>
   <p>Testing...</p>
   <div>normal <span>FADE</span> normal</div>
</body>
</html>
